v2.0.0-alpha feat: Add core functionality for BCG Screenshot Processor
- Implemented configuration constants in config.py for default settings and supported formats.
- Developed main entry point in main.py to handle directory input and initiate event processing.
- Created src/__init__.py to define the source package for the project.
- Added coverage analysis functionality in coverage_analyzer.py to check event coverage against ZIP files.
- Built MultiServerEventProcessor class in event_processor.py to manage event processing across multiple servers.
- Introduced file_utils.py for utility functions related to file handling and ZIP extraction.
- Developed excel_report.py to generate Excel reports with clickable links to screenshots.
- Implemented summary_generator.py to display a final summary of event categories processed.
- Added video_processing.py to handle video file operations, including screenshot extraction.
- Enhanced error handling and user prompts throughout the processing workflow.
