+++ src/__init__.py
@@ -1,0 +1,6 @@
+"""
+BCG Screenshot Processor - Source Package
+
+This package contains the core functionality for processing security event reports,
+extracting screenshots from videos, and generating Excel reports.
+"""
